+++ frontend/src/components/Table/Table.types.ts
@@ -1,0 +1,59 @@
+import { AdditionalDetails } from '../../types';
+
+export type AnimeRow = {
+  title: string;
+  genres: AdditionalDetails[];
+  type: string;
+  episodes: number;
+  airedFrom: string;
+  rating: string;
+  airedTo: string;
+  score: number;
+  id: number;
+};
+
+export type HeaderCell = {
+  label: string;
+  key: string;
+  isSortable: boolean;
+  sortDirection?: 'asc' | 'desc';
+  size: number;
+};
+
+export type ActiveSorting = {
+  columnId: string;
+  sortDirection: string;
+};
+
+export type TableProps = {
+  rows: AnimeRow[];
+  fetchNextPage: () => {};
+  isFetching: boolean;
+  headers: HeaderCell[];
+  activeSorting: ActiveSorting;
+  onSort: (_columnId: string, _sortDirection: string) => void;
+  hasError?: boolean
+};
+
+export type TableBodyRowProps = {
+  row: AnimeRow;
+};
+
+export type TableBodyProps = Pick<TableProps, 'rows' | 'fetchNextPage' | 'isFetching'>;
+
+export type TableHeaderCellProps = Pick<TableProps, 'onSort' | 'activeSorting'> & {
+  id: string;
+  value: string;
+  width?: number;
+  isSortable: boolean;
+};
+
+export type TableHeaderProps = Pick<TableProps, 'headers' | 'activeSorting' | 'onSort'>;
+
+export type TabelCellValue = string | number | AdditionalDetails[];
+
+export type TableCellProps = {
+  value: TabelCellValue;
+  type?: string;
+  width?: number;
+};
