+++ frontend/src/constants.ts
@@ -1,0 +1,28 @@
+export const HEADERS = [
+  {
+    label: 'Title', key: 'title', isSortable: true, size: (3 / 9) * 100,
+  },
+  {
+    label: 'Genre', key: 'genres', isSortable: false, size: (1 / 9) * 100,
+  },
+  {
+    label: 'Type', key: 'type', isSortable: false, size: (1 / 9) * 100,
+  },
+  {
+    label: 'Episodes', key: 'episodes', isSortable: true, size: (1 / 9) * 100,
+  },
+  {
+    label: 'Aired from', key: 'aired_from', isSortable: false, size: (1 / 9) * 100,
+  },
+  {
+    label: 'Aired to', key: 'aired_to', isSortable: false, size: (1 / 9) * 100,
+  },
+  {
+    label: 'Rating', key: 'rating', isSortable: false, size: (1 / 9) * 100,
+  },
+  {
+    label: 'Score', key: 'score', isSortable: true, size: (1 / 9) * 100,
+  },
+];
+
+export const DATE_FORMAT = 'yyyy/MM/dd';
