+++ frontend/src/containers/NotFound/NotFound.tsx
@@ -2,8 +2,9 @@ import React, { FC } from 'react';
 import { Link } from 'react-router';
 
 import { NotFoundContainer } from './NotFound.styled';
+import { NotFoundProps } from './NotFound.types';
 
-export const NotFound: FC<({ title?: string })> = ({ title = 'Ooops! Resource not found' }) => {
+export const NotFound: FC<(NotFoundProps)> = ({ title = 'Ooops! Resource not found' }) => {
   return (
     <NotFoundContainer>
       <h2>{title}</h2>
