+++ frontend/src/containers/NotFound/NotFound.types.ts
@@ -1,0 +1,3 @@
+export type NotFoundProps = {
+  title?: string;
+};
