+++ frontend/src/utils/flattenDataArray.ts
@@ -1,0 +1,8 @@
+import { AnimeItem } from '../types';
+
+export const flattenDataArray = (array: { data: AnimeItem[]; pagination: object }[]): any[] => array?.reduce((acc, item) => {
+  if (array && Array.isArray(item.data)) {
+    acc.push(...item.data);
+  }
+  return acc;
+}, [] as any[]);
